feat(metrics): Add rds_certificate_expiry_timestamp_seconds
This metric exposes the Unix timestamp of the expiration date of the instance certificate.
This is different from the expiration of the root certificate of the chain, which may be in more than 10 years in the future.

According to AWS documentation on https://docs.aws.amazon.com/AmazonRDS/latest/UserGuide/UsingWithRDS.SSL-certificate-rotation.html ,
the certificate for instances that support certificate renewal without restart, such as RDS and Aurora, is renewed at half time.
For those instances, the certificates is issued for 12 months.

Hence, we should expect a new expiry timestamp every ~6 months, with a value not lower than now minus 6 months.