fix: Add missing logger in servicequotas fetcher
This commit is updating the exporter and servicequotas packages to
properly pass the logger.

In the `serviceQuotaFetcher` struct we have defined a logger field but
it was never initialized. Which means any call to s.logger was panicking
in the `serviceQuotaFetcher` methods.

Properly passing the logger and storing it in the structure will allow
us to properly use `s.logger` inside the `serviceQuotaFetcher` methods.

This change has been made following this [issue][1]

[1]: https://github.com/qonto/prometheus-rds-exporter/issues/205

Signed-off-by: Clovis Delarue <clovis.delarue@qonto.com>