chore: use aws.ToString to avoid panic
This commit is updating a dereference of a string pointer to use
aws.ToString instead.

Aws is not really clear in their documentation if this
result.Quota.ErrorReason.ErrorMessage is never empty. To be sure we
don't panic if that's the case it's safer to use aws.ToString which will
dereference the pointer if it's not empty, and use an empty string if
the pointer is nil.

Signed-off-by: Clovis Delarue <clovis.delarue@qonto.com>