chore(config): Add parameter to enable log size collection for serverless instances
Collecting logs through the AWS RDS API for a serverless RDS instance requires the instance to be running. This prevents the instance from scaling down to zero and can result in unexpected costs when monitoring is enabled.

To address this, a new configurable parameter is introduced. By default, the exporter will not fetch logs for serverless instances to avoid triggering them to start, but it can be enabled for deployments that do not require scaling to zero.

Signed-off-by: Vincent Mercier <vincmer@amazon.com>